+++ paper/src/main/java/org/empirewar/lemonsqueezer/paper/LemonSqueezerPaper.java
@@ -1,0 +1,19 @@
+package org.empirewar.lemonsqueezer.paper;
+
+import org.bukkit.Bukkit;
+import org.bukkit.event.Listener;
+import org.bukkit.plugin.java.JavaPlugin;
+
+public final class LemonSqueezerPaper extends JavaPlugin implements Listener {
+
+    @Override
+    public void onEnable() {
+        Bukkit.getPluginManager().registerEvents(this, this);
+    }
+
+    @Override
+    public void onDisable() {
+        // Plugin shutdown logic
+    }
+
+}
