--- sponge/src/main/java/org/empirewar/lemonsqueeezer/LemonSqueezerSponge.java
@@ -1,3 +1,0 @@
-package org.empirewar.lemonsqueeezer;
-
-public class LemonSqueezerSponge {}
